// src/plugins/loyalty-points/gql/generated.ts
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  JSON: { input: any; output: any; }
  Money: { input: number; output: number; }
  Upload: { input: any; output: any; }
};
// ...
export type LoyaltyPointsTransactionInput = {
  customerId: Scalars['ID']['input'];
  note: Scalars['String']['input'];
  value: Scalars['Int']['input'];
};
// ...
export enum LoyaltyTransactionType {
  ADMINISTRATOR_ADJUSTED = 'ADMINISTRATOR_ADJUSTED',
  EARNED_ON_ORDER = 'EARNED_ON_ORDER',
  ORDER_CANCELLED = 'ORDER_CANCELLED',
  OTHER = 'OTHER',
  USED_ON_ORDER = 'USED_ON_ORDER'
}